update feature normalization for generation 2
