+++ nugraph/nugraph/util/__init__.py
@@ -3,7 +3,7 @@ from .RecallLoss import RecallLoss
 from .LogCoshLoss import LogCoshLoss
 from .ObjCondensationLoss import ObjCondensationLoss
 from .position_features import PositionFeatures
-from .FeatureNorm import FeatureNorm, FeatureNormMetric
+from .feature_norm import FeatureNorm, FeatureNormMetric
 from .hierarchical_edges import HierarchicalEdges
 from .event_labels import EventLabels
 from .scriptutils import setup_env, configure_device
